refactor deskew
