+++ biahub/cli/main.py
@@ -3,7 +3,6 @@ import click
 from biahub.cli.characterize_psf import characterize_psf
 from biahub.cli.concatenate import concatenate_cli
 from biahub.cli.deconvolve import deconvolve
-from biahub.cli.deskew import deskew
 from biahub.cli.estimate_bleaching import estimate_bleaching
 from biahub.cli.estimate_deskew import estimate_deskew
 from biahub.cli.estimate_psf import estimate_psf
@@ -15,6 +14,7 @@ from biahub.cli.register import register
 from biahub.cli.segment import segment
 from biahub.cli.stabilize import stabilize
 from biahub.cli.stitch import stitch
+from biahub.deskew import deskew_cli
 
 CONTEXT = {"help_option_names": ["-h", "--help"]}
 
@@ -32,7 +32,7 @@ def cli():
 
 cli.add_command(estimate_bleaching)
 cli.add_command(estimate_deskew)
-cli.add_command(deskew)
+cli.add_command(deskew_cli)
 cli.add_command(estimate_registration)
 cli.add_command(optimize_registration)
 cli.add_command(register)
